requirements.txt
langgraph
faiss-cpu
sentence-transformers
numpy
scikit-learn